// apps/web/src/types/nav.ts
import { type Icons } from '@/components/icons';

export type NavItem = {
  title: string;
  href?: string;
  disabled?: boolean;
  external?: boolean;
  icon?: keyof typeof Icons;
  label?: string;
};

export type NavItemWithChildren = {
  items: NavItemWithChildren[];
} & NavItem;

export type MainNavItem = {} & NavItem;

export type SidebarNavItem = {} & NavItemWithChildren;